Increase font size for mobile
Note to self: the reason it looks like the sidebar shifts a little when
changing to certain pages is because the browser adds the scrollbar.
This means, this behavior only occurs when clicking from a page with a
scrollbar to one without (usually the home page to something else, like
my reading log). I guess the only way to fix this is by writing more
posts so the homepage has a scrollbar :)